+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/calculator/AssetAllocationCalcResult.java
@@ -2,13 +2,27 @@ package com.strangegrotto.wealthdraft.assetallocation.calculator;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
+import com.strangegrotto.wealthdraft.assets.definition.Asset;
 import org.immutables.value.Value;
+
+import java.math.BigDecimal;
 
 @WealthdraftImmutableStyle
 @Value.Immutable
-@JsonDeserialize(as = ImmAssetAllocationCalcResults.class)
 public interface AssetAllocationCalcResult {
-    getNumeratorFilterName();
+    BigDecimal getCurrentNumeratorValue();
 
-    getDenominatorFilterName();
+    BigDecimal getCurrentDenominatorValue();
+
+    BigDecimal getCurrentFraction();
+
+    BigDecimal getDesiredFraction();
+
+    BigDecimal getDesiredNumeratorValue();
+
+    BigDecimal getCorrectionNeeded();
+
+    BigDecimal getDeviationFraction();
+
+    AssetAllocationDeviationStatus getDeviationStatus();
 }