+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/calculator/AssetAllocationCalculator.java
@@ -1,38 +1,29 @@
 package com.strangegrotto.wealthdraft.assetallocation.calculator;
 
-import com.jakewharton.picnic.CellStyle;
-import com.jakewharton.picnic.Table;
-import com.jakewharton.picnic.TableSection;
-import com.jakewharton.picnic.TextAlignment;
-import com.strangegrotto.wealthdraft.Display;
-import com.strangegrotto.wealthdraft.assetallocation.AssetAllocationRenderer;
-import com.strangegrotto.wealthdraft.assetallocation.TargetAssetAllocations;
-import com.strangegrotto.wealthdraft.assetallocation.filters.AssetFilter;
+import com.strangegrotto.wealthdraft.assetallocation.datamodel.TargetAssetAllocation;
+import com.strangegrotto.wealthdraft.assetallocation.datamodel.TargetAssetAllocations;
+import com.strangegrotto.wealthdraft.assetallocation.datamodel.filters.AssetFilter;
 import com.strangegrotto.wealthdraft.assets.definition.Asset;
 import com.strangegrotto.wealthdraft.assets.temporal.AssetSnapshot;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 public class AssetAllocationCalculator {
+    // The rounding parameters to use so that non-terminating divison doesn't throw an exception
+    private static final int DIVISION_SCALE = 6;
     private static final RoundingMode ROUNDING_MODE = RoundingMode.HALF_EVEN;
-    private static final int BIG_DECIMAL_DISPLAY_SCALE = 2;  // TODO replace this with something on Display
 
-    // TODO Use these to color the output accordingly
-    private final double deviationPercentageWarn;
-    private final double deviationPercentageError;
+    private final BigDecimal deviationFractionWarn;
+    private final BigDecimal deviationFractionErr;
 
-    public AssetAllocationRenderer(double deviationPercentageWarn, double deviationPercentageError) {
-        this.deviationPercentageWarn = deviationPercentageWarn;
-        this.deviationPercentageError = deviationPercentageError;
+    public AssetAllocationCalculator(BigDecimal deviationFractionWarn, BigDecimal deviationFractionErr) {
+        this.deviationFractionWarn = deviationFractionWarn;
+        this.deviationFractionErr = deviationFractionErr;
     }
 
-    public List<AssetAllocationCalcResult> calculateAssetAllocations(
+    public LinkedHashMap<TargetAssetAllocation, AssetAllocationCalcResult> calculate(
             TargetAssetAllocations targetAssetAllocations,
             Map<String, Asset<?, ?>> assets,
             Map<String, AssetSnapshot<?>> latestAssetSnapshots) {
@@ -43,8 +34,7 @@ public class AssetAllocationCalculator {
         var filters = targetAssetAllocations.getFilters();
         var targets = targetAssetAllocations.getTargets();
 
-        // GET TO DAH CHOPPAH!
-        var results = new ArrayList<AssetAllocationCalcResult();
+        var results = new LinkedHashMap<TargetAssetAllocation, AssetAllocationCalcResult>();
         for (var target : targets) {
             var numeratorFilterName = target.getNumeratorFilter();
             var numeratorFilter = filters.get(numeratorFilterName);
@@ -63,34 +53,41 @@ public class AssetAllocationCalculator {
                 denominatorStrRepr = "Total Portfolio";
             }
 
-            var calcResult = ImmAssetAllocationCalcResults.of(
-                    numeratorFilterName,
-                    denominatorStrRepr,
+            var currentFraction = numeratorValue.divide(
+                    denominatorValue,
+                    DIVISION_SCALE,
+                    ROUNDING_MODE
+            );
+            var targetFraction = target.getFraction();
+            var targetNumeratorValue = targetFraction.multiply(denominatorValue);
+            var correctionNeeded = targetNumeratorValue.subtract(numeratorValue);
+            var deviationFraction = correctionNeeded.abs().divide(
+                    targetNumeratorValue,
+                    DIVISION_SCALE,
+                    ROUNDING_MODE
+            );
+            AssetAllocationDeviationStatus deviationStatus;
+            if (deviationFraction.compareTo(this.deviationFractionErr) >= 0) {
+                deviationStatus = AssetAllocationDeviationStatus.ERROR;
+            } else if (deviationFraction.compareTo(this.deviationFractionWarn) >= 0) {
+                deviationStatus = AssetAllocationDeviationStatus.WARN;
+            } else {
+                deviationStatus = AssetAllocationDeviationStatus.OK;
+            }
 
-            )
-            results.add()
-            addTableRow(
-                    tableBodyBuilder,
-                    numeratorFilterName,
-                    denominatorStrRepr,
+            var calcResult = ImmAssetAllocationCalcResult.of(
                     numeratorValue,
                     denominatorValue,
-                    target.getFraction()
+                    currentFraction,
+                    targetFraction,
+                    targetNumeratorValue,
+                    correctionNeeded,
+                    deviationFraction,
+                    deviationStatus
             );
+            results.put(target, calcResult);
         }
-
-        // TODO Replace this with a better method inside Display
-        var tableBody = tableBodyBuilder.build();
-        var cellStyle = new CellStyle.Builder()
-                .setAlignment(TextAlignment.MiddleCenter)
-                .setPadding(1)
-                .setBorder(true)
-                .build();
-        var table = new Table.Builder()
-                .setCellStyle(cellStyle)
-                .setBody(tableBody)
-                .build();
-        System.out.println(table.toString());
+        return results;
     }
 
     private static BigDecimal getValueOfAssetsMatchingFilter(
@@ -103,38 +100,5 @@ public class AssetAllocationCalculator {
                 .map(Map.Entry::getValue)
                 .map(AssetSnapshot::getValue)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
-    }
-
-    private static void addTableRow(
-            TableSection.Builder tableBodyBuilder,
-            String numeratorSelectors,
-            String denominatorSelectors,
-            BigDecimal numeratorValue,
-            BigDecimal denominatorValue,
-            BigDecimal targetFraction) {
-        var currentFraction = numeratorValue.divide(
-                denominatorValue,
-                BIG_DECIMAL_DISPLAY_SCALE + 2,  // Plus 2 so that when we multiply by 100 to get a fraction we have scale == 2
-                RoundingMode.HALF_EVEN);
-        var currentPercent = formatFractionAsPercent(currentFraction);
-        var targetPercent = formatFractionAsPercent(targetFraction);
-        var targetValue = targetFraction.multiply(denominatorValue).setScale(BIG_DECIMAL_DISPLAY_SCALE, RoundingMode.HALF_EVEN);
-        var correctionNeeded = targetValue.subtract(numeratorValue).setScale(BIG_DECIMAL_DISPLAY_SCALE, RoundingMode.HALF_EVEN);
-
-        // TODO Pull back a preferred currency format and use that here instead!
-        var strRow = List.of(
-                numeratorSelectors,
-                denominatorSelectors,
-                currentPercent.toString(),
-                targetPercent.toString(),
-                correctionNeeded.toString()
-        );
-        var strRowArr = strRow.toArray(new String[0]);
-        tableBodyBuilder.addRow(strRowArr);
-    }
-
-    private static BigDecimal formatFractionAsPercent(BigDecimal input) {
-        var scaledUp = input.multiply(BigDecimal.valueOf(100));
-        return scaledUp.setScale(BIG_DECIMAL_DISPLAY_SCALE, ROUNDING_MODE);
     }
 }