+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/calculator/AssetAllocationDeviationStatus.java
@@ -1,6 +1,6 @@
 package com.strangegrotto.wealthdraft.assetallocation.calculator;
 
-public enum AssetAllocationStatus {
+public enum AssetAllocationDeviationStatus {
     OK,     // Asset allocation deviation is within tolerances
     WARN,   // Asset allocation % has deviated somewhat
     ERROR,  // Asset allocation % has deviated significantly