+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/NegationAssetFilter.java
@@ -1,8 +1,7 @@
-package com.strangegrotto.wealthdraft.assetallocation.filters;
+package com.strangegrotto.wealthdraft.assetallocation.datamodel.filters;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.google.common.collect.Sets;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
 import com.strangegrotto.wealthdraft.assets.definition.Asset;
 import org.immutables.value.Value;