+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/calculator/AssetAllocationCalculatorTest.java
@@ -1,0 +1,5 @@
+package com.strangegrotto.wealthdraft.assetallocation.calculator;
+
+public class AssetAllocationCalculatorTest {
+    // TODO
+}